+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [10.1.5](https://github.com/Doist/todoist-sdk-typescript/compare/v10.1.4...v10.1.5) (2026-05-13)
+
+### Bug Fixes
+
+* read attachment/backup bodies as binary in customFetch branch ([#602](https://github.com/Doist/todoist-sdk-typescript/issues/602)) ([313e7d6](https://github.com/Doist/todoist-sdk-typescript/commit/313e7d641f4ceb2b66c9551e51c74bfe5c0cad6d))
+
+## [10.1.4](https://github.com/Doist/todoist-sdk-typescript/compare/v10.1.3...v10.1.4) (2026-05-13)
+
+### Bug Fixes
+
+* Convert `uploadFile` response from snake_case to camelCase before validating ([#601](https://github.com/Doist/todoist-sdk-typescript/issues/601)) ([7d0c038](https://github.com/Doist/todoist-sdk-typescript/commit/7d0c0387f2ebb7b24d1a8e7debe3b184a857d015))
+
 ## [10.1.3](https://github.com/Doist/todoist-sdk-typescript/compare/v10.1.2...v10.1.3) (2026-05-09)
 
 ### Bug Fixes